more and some refactoring
