+++ src/app/virtualmachines/clusterdetail/clusterdetail.component.html
@@ -47,6 +47,14 @@
           </div>
         </div>
       </div>
+      <div class="row">
+         <div class="col-3">
+          <div class="callout">
+            <small class="text-muted">Public IP</small><br>
+            <strong class="text-muted">{{cluster?.public_ip}}</strong>
+          </div>
+        </div>
+      </div>
     </div>
     <hr>
