--- src/app/virtualmachines/virtualmachineinfo/virtualmachineinfo.component.html
@@ -1,84 +1,0 @@
-<div class="row">
-  <div class="col-4" style="margin:auto">
-    <div [ngSwitch]="vm.status">
-      <div *ngSwitchCase="VirtualMachineStates.staticACTIVE"
-           [id]="'active_badge_' + vm.name"
-           class="dot-colorless bg-success pulseAnim" style="width: 15px; height: 15px;">
-      </div>
-      <div *ngSwitchCase="VirtualMachineStates.staticSHUTOFF"
-           [id]="'shutoff_badge_' + vm.name"
-           class="dot-colorless bg-warning" style="width: 15px; height: 15px">
-      </div>
-      <div *ngSwitchCase="VirtualMachineStates.staticDELETED"
-           [id]="'deleted_badge_' + vm.name"
-           class="dot-colorless bg-danger" style="width: 15px; height: 15px">
-      </div>
-      <div *ngSwitchCase="VirtualMachineStates.staticBUILD"
-           [id]="'build_badge_' + vm.name"
-           class="dot-colorless bg-dark" style="width: 15px; height: 15px">
-      </div>
-      <div *ngSwitchCase="VirtualMachineStates.staticPREPARE_PLAYBOOK_BUILD"
-           [id]="'ppb_badge_' + vm.name"
-           class="dot-colorless bg-dark" style="width: 15px; height: 15px">
-      </div>
-      <div *ngSwitchCase="VirtualMachineStates.staticBUILD_PLAYBOOK"
-           [id]="'build_playbook_badge_' + vm.name"
-           class="dot-colorless bg-dark" style="width: 15px; height: 15px">
-      </div>
-      <div *ngSwitchCase="VirtualMachineStates.staticRESTARTING"
-           [id]="'restarting_badge_' + vm.name"
-           class="dot-colorless bg-dark" style="width: 15px; height: 15px">
-      </div>
-      <div *ngSwitchCase="VirtualMachineStates.staticPOWERING_OFF"
-           [id]="'powering_off_badge_' + vm.name"
-           class="dot-colorless bg-dark" style="width: 15px; height: 15px">
-      </div>
-      <div *ngSwitchDefault
-           [id]="'default_badge_' + vm.name"
-           class="dot-colorless bg-dark" style="width: 15px; height: 15px">
-      </div>
-      <span style="margin-left:15px;"
-            class="text-muted font-weight-bold">Status: {{vm.status}}</span>
-    </div>
-  </div>
-  <div class="col-2">
-    <div class="callout callout-success">
-      <small class="text-muted">Created at (y-m-d)</small><br>
-      <strong class="text-muted">{{vm.created_at | date: 'yyyy.MM.dd HH:mm'}}</strong>
-    </div>
-  </div>
-  <div class="col-2"
-       *ngIf="vm.stopped_at && vm.status === VirtualMachineStates.staticSHUTOFF">
-    <div class="callout callout-warning">
-      <small class="text-muted">Stopped at (y-m-d)</small><br>
-      <strong class="text-muted">{{vm.stopped_at | date: 'yyyy.MM.dd HH:mm'}}</strong>
-    </div>
-  </div>
-  <div class="col-2">
-    <div class="callout">
-      <small class="text-muted">Part of</small><br>
-      <a [routerLink]="['/project-management/' + vm.application_id]" target="_blank">
-        <strong class="text-muted">{{vm.project}}</strong></a>
-    </div>
-  </div>
-  <div class="col-4">
-    <div class="callout">
-      <small class="text-muted">Openstack-ID</small><br>
-      <strong class="text-muted">{{vm.openstackid}}</strong>
-    </div>
-  </div>
-</div>
-<div class="row">
-  <div class="col-4">
-    <div class="callout">
-      <small class="text-muted">Created by</small><br>
-      <strong class="text-muted">{{vm.userlogin}}</strong>
-    </div>
-  </div>
-  <div class="col-8">
-    <div class="callout">
-      <small class="text-muted">Elixir-ID of {{vm.userlogin}}</small><br>
-      <strong class="text-muted">{{vm.elixir_id}}</strong>
-    </div>
-  </div>
-</div>
