--- src/app/virtualmachines/vmdetail.module.ts
@@ -1,29 +1,0 @@
-import {NgModule} from '@angular/core';
-import {AccordionModule} from 'ngx-bootstrap';
-import {TabsModule} from 'ngx-bootstrap/tabs';
-import {FormsModule} from '@angular/forms';
-import {CommonModule} from '@angular/common';
-import {ModalModule} from 'ngx-bootstrap/modal';
-import {SharedDirectivesModule} from '../shared/shared_modules/shared_directives.module';
-import {VmDetailRoutingModule} from './vmdetail-routing.module';
-
-/**
- * Vm Detail Module
- */
-@NgModule({
-  imports: [
-    AccordionModule.forRoot(),
-    VmDetailRoutingModule,
-    TabsModule,
-    FormsModule,
-    CommonModule,
-    ModalModule.forRoot(),
-    SharedDirectivesModule
-  ],
-  declarations: [
-  ]
-})
-
-export class VmDetailModule {
-
-}
