+++ src/app/virtualmachines/vmdetail/virtualmachineinfo/virtualmachineinfo.component.ts
@@ -1,6 +1,6 @@
 import {Component, Input, OnInit} from '@angular/core';
-import {VirtualMachineStates} from '../virtualmachinemodels/virtualmachinestates';
-import {VirtualMachine} from '../virtualmachinemodels/virtualmachine';
+import {VirtualMachineStates} from '../../virtualmachinemodels/virtualmachinestates';
+import {VirtualMachine} from '../../virtualmachinemodels/virtualmachine';
 
 @Component({
              selector: 'app-virtualmachineinfo',
