+++ src/app/virtualmachines/vmdetail/vmdetail.component.html
